remove global usage by adding a new context for the meteor config init and import
