+++ tools/project-context.js
@@ -1,3 +1,4 @@
+import { normalizeModernConfig, setMeteorConfig } from "./tool-env/meteor-config";
 
 var assert = require("assert");
 var _ = require('underscore');
@@ -1769,26 +1770,6 @@ Object.assign(exports.FinishedUpgraders.prototype, {
   }
 });
 
-const DEFAULT_MODERN = {
-  transpiler: true,
-  minifier: true,
-  webArchOnly: true,
-  watcher: true,
-};
-
-export const normalizeModern = (r = false) => Object.fromEntries(
-    Object.entries(DEFAULT_MODERN).map(([k, def]) => [
-      k,
-      r === true
-          ? def
-          : r === false || r?.[k] === false
-              ? false
-              : typeof r?.[k] === 'object'
-                  ? { ...r[k] }
-                  : def,
-    ]),
-);
-
 export class MeteorConfig {
   constructor({
     appDirectory,
@@ -1836,10 +1817,11 @@ export class MeteorConfig {
     const modernForced = JSON.parse(process.env.METEOR_MODERN || "false");
     // Reinitialize meteorConfig globally for project context
     // Updates config when package.json changes trigger rebuilds
-    global.meteorConfig = {
+    setMeteorConfig({
       ...(this._config || {}),
-      modern: normalizeModern(modernForced || this._config?.modern || false),
-    };
+      modern: normalizeModernConfig(modernForced || this._config?.modern || false),
+    });
+
     return this._config;
   }
